[Add]通知機能一覧
