+++ app/assets/javascripts/notifications.coffee
