+++ app/controllers/comments_controller.rb
@@ -4,9 +4,11 @@ class CommentsController < ApplicationController
 
     def create
     @comment = Comment.new(comment_params)
+    @comment.user_id = current_user.id
     @comment.save
-    @post.create_notification_comment!(current_user, @comment.id)
+
     @task = @comment.task
+    @task.create_notification_comment!(current_user, @comment.id)
     end
 
   def destroy
@@ -18,6 +20,6 @@ class CommentsController < ApplicationController
   private
 
   def comment_params
-    params.require(:comment).permit(:comment, :task_id, :user_id)
+    params.require(:comment).permit(:comment, :task_id)
   end
 end
