--- app/helpers/notification_helper.rb
@@ -1,2 +1,0 @@
-module NotificationHelper
-end
