+++ app/helpers/notifications_helper.rb
@@ -1,0 +1,2 @@
+module NotificationsHelper
+end
