+++ app/models/task.rb
@@ -1,14 +1,15 @@
 class Task < ApplicationRecord
   belongs_to :user
   has_many :comments, dependent: :destroy
+  has_many :notifications, dependent: :destroy
 
   enum task_status: {未着手: 0, 進行中: 1, 保留: 2, 確認中: 3, 遅れ: 4, 完了: 5}
   enum task_matrix: {重要かつ緊急: 0, 重要で緊急でない: 1, 重要でなく緊急: 2, 重要でないかつ緊急でない: 3 }
-  
-  
+
+
   def create_notification_comment!(current_user, comment_id)
     # 自分以外にコメントしている人をすべて取得し、全員に通知を送る
-    temp_ids = Comment.select(:user_id).where(post_id: id).where.not(user_id: current_user.id).distinct
+    temp_ids = Comment.select(:user_id).where(task_id: id).where.not(user_id: current_user.id).distinct
     temp_ids.each do |temp_id|
       save_notification_comment!(current_user, comment_id, temp_id["user_id"])
     end
