+++ db/migrate/20210125012641_create_notifications.rb
@@ -4,6 +4,7 @@ class CreateNotifications < ActiveRecord::Migration[5.2]
       t.integer :visitor_id
       t.integer :visited_id
       t.integer :comment_id
+      t.integer :task_id
       t.string :action, default: "", null: false
       t.boolean :checked, default: false, null: false
