+++ db/schema.rb
@@ -24,6 +24,7 @@ ActiveRecord::Schema.define(version: 2021_01_25_012641) do
     t.integer "visitor_id"
     t.integer "visited_id"
     t.integer "comment_id"
+    t.integer "task_id"
     t.string "action", default: "", null: false
     t.boolean "checked", default: false, null: false
     t.datetime "created_at", null: false
