+++ test/controllers/notifications_controller_test.rb
@@ -1,0 +1,9 @@
+require 'test_helper'
+
+class NotificationsControllerTest < ActionDispatch::IntegrationTest
+  test "should get index" do
+    get notifications_index_url
+    assert_response :success
+  end
+
+end
